backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.2.0
# Phase 3: Virtual Try-On Engine
numpy>=1.26.0
opencv-python-headless>=4.9.0
rembg>=2.0.55
google-genai>=1.0.0
# Phase 1 Infrastructure & Phase 2 Security
motor==3.3.2
aioredis==2.0.1
boto3==1.34.0
pymongo==4.6.0
certifi
slowapi==0.1.9
bleach==6.1.0
